Refactor client code by extracting connection logic to separate entities
- + fix typo in FileDownloader notification

+++ hw-03-torrent/client/src/main/java/ru/itmo/javacourse/torrent/Main.java
@@ -14,7 +14,7 @@ public class Main {
         if (args.length != 1) {
             System.out.println("Specify port number");
         }
-        final int port = Integer.valueOf(args[0]);
+        final short port = Short.valueOf(args[0]);
         Scanner scanner = new Scanner(System.in);
         Cli cli = new Cli();
         List<File> filesToUpload = cli.addFilesToUploadPrompt(scanner);
